Move compile command test
