+++ docs/source/tests/python/utils.rst
@@ -6,6 +6,11 @@ Utils
 
 .. automodule:: tests.python.utils.test_cmake
 
+Compile command
+~~~~~~~~~~~~~~~
+
+.. automodule:: tests.python.utils.test_compile_command
+
 Detect GPUs
 ~~~~~~~~~~~
