+++ tests/python/utils/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_pytest(cmake)
+add_pytest(compile_command)
 add_pytest(detect)
